Create a handle_helpers.go file
